refactor module user
- move dtos to service directory
- update returned type is a pointer
