+++ modules/user/infras/controller/http-gin/authenticate_api.go
@@ -4,12 +4,12 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	usermodel "github.com/ntttrang/go-food-delivery-backend-service/modules/user/model"
+	service "github.com/ntttrang/go-food-delivery-backend-service/modules/user/service"
 	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
 )
 
 func (ctrl *UserHttpController) AuthenticateAPI(c *gin.Context) {
-	var req usermodel.AuthenticateReq
+	var req service.AuthenticateReq
 	if err := c.ShouldBindJSON(&req); err != nil {
 		panic(datatype.ErrBadRequest.WithError(err.Error()))
 	}
