+++ modules/user/infras/controller/http-gin/create_api.go
@@ -4,12 +4,12 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	usermodel "github.com/ntttrang/go-food-delivery-backend-service/modules/user/model"
+	service "github.com/ntttrang/go-food-delivery-backend-service/modules/user/service"
 	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
 )
 
 func (ctrl *UserHttpController) CreateUserAPI(c *gin.Context) {
-	var requestBodyData *usermodel.CreateUserReq
+	var requestBodyData *service.CreateUserReq
 
 	if err := c.ShouldBindJSON(&requestBodyData); err != nil {
 		panic(datatype.ErrBadRequest.WithError(err.Error()))
