+++ modules/user/infras/repository/gorm-mysql/find.go
@@ -1,4 +1,4 @@
-package userrepository
+package usergormmysql
 
 import (
 	"context"
