+++ modules/user/infras/repository/gorm-mysql/insert_user_address.go
@@ -1,4 +1,4 @@
-package userrepository
+package usergormmysql
 
 import (
 	"context"
