+++ modules/user/infras/repository/gorm-mysql/repo.go
@@ -1,4 +1,4 @@
-package userrepository
+package usergormmysql
 
 import (
 	shareinfras "github.com/ntttrang/go-food-delivery-backend-service/shared/infras"
