--- modules/user/model/authenticate_res.go
@@ -1,6 +1,0 @@
-package usermodel
-
-type AuthenticateRes struct {
-	Token string `json:"token"`
-	ExpIn int    `json:"expIn"`
-}
