--- modules/user/model/introspect_req.go
@@ -1,15 +1,0 @@
-package usermodel
-
-import "errors"
-
-type IntrospectReq struct {
-	Token string `json:"token"`
-}
-
-func (c *IntrospectReq) Validate() error {
-	if c.Token == "" {
-		return errors.New("token is required")
-	}
-
-	return nil
-}
