--- modules/user/model/user_detail_req.go
@@ -1,7 +1,0 @@
-package usermodel
-
-import "github.com/google/uuid"
-
-type UserDetailReq struct {
-	Id uuid.UUID `json:"id"`
-}
