--- modules/user/model/user_list_req.go
@@ -1,9 +1,0 @@
-package usermodel
-
-import sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
-
-type UserListReq struct {
-	UserSearchDto
-	sharedModel.PagingDto
-	sharedModel.SortingDto
-}
