--- modules/user/model/user_search_dto.go
@@ -1,8 +1,0 @@
-package usermodel
-
-type UserSearchDto struct {
-	Name  string `json:"name"`
-	Email string `json:"email"`
-	Phone string `json:"phone"`
-	Role  string `json:"role"`
-}
