+++ modules/user/service/create_user.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"strings"
 	"time"
 
 	"github.com/google/uuid"
@@ -10,6 +11,63 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type CreateUserReq struct {
+	Email     string                `json:"email"`
+	Password  string                `json:"password"`
+	FirstName string                `json:"firstName"`
+	LastName  string                `json:"lastName"`
+	Role      *usermodel.UserRole   `json:"role"`
+	Type      *usermodel.UserType   `json:"userType"`
+	Status    *usermodel.UserStatus `json:"status"`
+	Phone     *string               `json:"phone"`
+
+	Id uuid.UUID `json:"-"`
+}
+
+func (r *CreateUserReq) Validate() error {
+	r.Email = strings.TrimSpace(r.Email)
+	r.Password = strings.TrimSpace(r.Password)
+	r.FirstName = strings.TrimSpace(r.FirstName)
+	r.LastName = strings.TrimSpace(r.LastName)
+
+	if r.Email == "" {
+		return usermodel.ErrEmailRequired
+	}
+
+	if !sharedModel.ValidateEmail(r.Email) {
+		return usermodel.ErrEmailInvalid
+	}
+
+	if len(r.Password) <= 6 {
+		return usermodel.ErrPasswordInvalid
+	}
+
+	if r.FirstName == "" {
+		return usermodel.ErrFirstNameRequired
+	}
+
+	if r.LastName == "" {
+		return usermodel.ErrLastNameRequired
+	}
+
+	return nil
+}
+
+func (r CreateUserReq) ConvertToUser() *usermodel.User {
+	return &usermodel.User{
+		Email:     r.Email,
+		Password:  r.Password,
+		FirstName: r.FirstName,
+		LastName:  r.LastName,
+		Role:      *r.Role,
+		Type:      *r.Type,
+		Status:    *r.Status,
+		Phone:     *r.Phone,
+	}
+}
+
+// Initilize service
 type ICreateRepo interface {
 	Insert(ctx context.Context, data *usermodel.User) error
 }
@@ -22,7 +80,8 @@ func NewCreateCommandHandler(userRepo ICreateRepo) *CreateCommandHandler {
 	return &CreateCommandHandler{userRepo: userRepo}
 }
 
-func (s *CreateCommandHandler) Execute(ctx context.Context, req *usermodel.CreateUserReq) error {
+// Implement
+func (s *CreateCommandHandler) Execute(ctx context.Context, req *CreateUserReq) error {
 	if err := req.Validate(); err != nil {
 		return datatype.ErrBadRequest.WithWrap(err).WithDebug(err.Error())
 	}
