+++ modules/user/service/generate_code.go
@@ -13,6 +13,18 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type Verification struct {
+	Id        uuid.UUID `gorm:"type:uuid;primary_key;default:gen_random_uuid()"`
+	UserId    uuid.UUID `gorm:"type:uuid;not null;index"`
+	Code      string    `gorm:"type:varchar(64);not null"`
+	ExpiresAt time.Time `gorm:"not null"`
+	Used      bool      `gorm:"not null;default:false"`
+	CreatedAt time.Time `gorm:"not null;default:current_timestamp"`
+	UpdatedAt time.Time `gorm:"not null;default:current_timestamp"`
+}
+
+// Initilize service
 type IUserRepo interface {
 	FindById(ctx context.Context, id uuid.UUID) (*usermodel.User, error)
 }
@@ -39,6 +51,7 @@ func NewGenerateCode(userRepo IUserRepo, redisCache IRedisCache, emailHdl IEmail
 	}
 }
 
+// Implement
 func (g *GenerateCode) Execute(ctx context.Context, userId uuid.UUID) (string, error) {
 
 	user, err := g.userRepo.FindById(ctx, userId)
