+++ modules/user/service/get_user_detail.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 	"errors"
+	"time"
 
 	"github.com/google/uuid"
 	"github.com/jinzhu/copier"
@@ -11,6 +12,23 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type UserDetailReq struct {
+	Id uuid.UUID `json:"id"`
+}
+
+type UserSearchResDto struct {
+	Id        uuid.UUID          `json:"id"`
+	FirstName string             `json:"firstName"`
+	LastName  string             `json:"lastName"`
+	Role      usermodel.UserRole `json:"role"`
+	Email     string             `json:"email"`
+	Phone     string             `json:"phone"`
+	CreatedAt *time.Time         `json:"createdAt"`
+	UpdatedAt *time.Time         `json:"updatedAt"`
+}
+
+// Initilize service
 type IGetDetailQueryRepo interface {
 	FindById(ctx context.Context, id uuid.UUID) (*usermodel.User, error)
 }
@@ -23,23 +41,24 @@ func NewGetDetailQueryHandler(repo IGetDetailQueryRepo) *GetDetailQueryHandler {
 	return &GetDetailQueryHandler{repo: repo}
 }
 
-func (hdl *GetDetailQueryHandler) Execute(ctx context.Context, req usermodel.UserDetailReq) (usermodel.UserSearchResDto, error) {
+// Implement
+func (hdl *GetDetailQueryHandler) Execute(ctx context.Context, req UserDetailReq) (UserSearchResDto, error) {
 	user, err := hdl.repo.FindById(ctx, req.Id)
 
 	if err != nil {
 		if errors.Is(err, usermodel.ErrUserNotFound) {
-			return usermodel.UserSearchResDto{}, datatype.ErrNotFound.WithDebug(usermodel.ErrUserNotFound.Error())
+			return UserSearchResDto{}, datatype.ErrNotFound.WithDebug(usermodel.ErrUserNotFound.Error())
 		}
-		return usermodel.UserSearchResDto{}, datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
+		return UserSearchResDto{}, datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
 	}
 
 	if user.Status == sharedModel.StatusDelete {
-		return usermodel.UserSearchResDto{}, datatype.ErrDeleted.WithError(usermodel.ErrUserDeletedOrBanned.Error())
+		return UserSearchResDto{}, datatype.ErrDeleted.WithError(usermodel.ErrUserDeletedOrBanned.Error())
 	}
 
-	var resp usermodel.UserSearchResDto
+	var resp UserSearchResDto
 	if err := copier.Copy(&resp, &user); err != nil {
-		return usermodel.UserSearchResDto{}, datatype.ErrInternalServerError.WithWrap(errors.New("copier libraries failed"))
+		return UserSearchResDto{}, datatype.ErrInternalServerError.WithWrap(errors.New("copier libraries failed"))
 	}
 	return resp, nil
 }
