+++ modules/user/service/update_user.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 	"errors"
+	"strings"
 
 	"github.com/google/uuid"
 	usermodel "github.com/ntttrang/go-food-delivery-backend-service/modules/user/model"
@@ -10,9 +11,41 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type UpdateUserReq struct {
+	FirstName string `json:"firstName"`
+	LastName  string `json:"lastName"`
+	Phone     string `json:"phone"`
+	Avatar    string `json:"avatar"`
+	Status    string `json:"status"`
+
+	Id uuid.UUID `json:"_"`
+}
+
+func (UpdateUserReq) TableName() string {
+	return usermodel.User{}.TableName()
+}
+
+func (r *UpdateUserReq) Validate() error {
+	r.FirstName = strings.TrimSpace(r.FirstName)
+	r.LastName = strings.TrimSpace(r.LastName)
+	r.Phone = strings.TrimSpace(r.Phone)
+
+	if r.Id.String() == "" {
+		return usermodel.ErrIdRequired
+	}
+
+	if r.Phone != "" && !sharedModel.ValidatePhoneNumber(r.Phone) {
+		return usermodel.ErrInvalidPhoneNumber
+	}
+
+	return nil
+}
+
+// Initilize service
 type IupdateRepo interface {
 	FindById(ctx context.Context, id uuid.UUID) (*usermodel.User, error)
-	Update(ctx context.Context, req *usermodel.UpdateUserReq) error
+	Update(ctx context.Context, req *UpdateUserReq) error
 }
 
 type UpdateCommandHandler struct {
@@ -23,7 +56,8 @@ func NewUpdateCommandHandler(userRepo IupdateRepo) *UpdateCommandHandler {
 	return &UpdateCommandHandler{userRepo: userRepo}
 }
 
-func (hdl *UpdateCommandHandler) Execute(ctx context.Context, req usermodel.UpdateUserReq) error {
+// Implement
+func (hdl *UpdateCommandHandler) Execute(ctx context.Context, req UpdateUserReq) error {
 	if err := req.Validate(); err != nil {
 		return datatype.ErrBadRequest.WithWrap(err).WithDebug(err.Error())
 	}
